split up PlayCard component
